requirements.txt
DBUtils==2.0
decorator==4.4.2
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.19.3
openpyxl==3.0.5
pandas==1.1.3
pbr==5.5.1
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.23
sqlalchemy-migrate==0.13.0
sqlparse==0.4.1
Tempita==0.5.2
